Wrap resource classes with event triggering
Removed private event publishers. Add wrapper classes to QueueHandler, Consumer, Queue, ExchangeRegistry, QueueRegistry to provide event triggering capability.
