+++ modules/broker-core/src/main/java/io/ballerina/messaging/broker/core/QueueRegistry.java
@@ -1,177 +1,32 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
 package io.ballerina.messaging.broker.core;
 
 import io.ballerina.messaging.broker.common.ResourceNotFoundException;
 import io.ballerina.messaging.broker.common.ValidationException;
 import io.ballerina.messaging.broker.common.data.types.FieldTable;
-import io.ballerina.messaging.broker.core.store.dao.QueueDao;
-import io.ballerina.messaging.broker.eventing.EventSync;
 
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
 
 /**
- * Registry object which contains all the queues of the broker.
+ * Abstract class for Queue Registry objects.
  */
-public final class QueueRegistry {
+public abstract class QueueRegistry {
 
-    private final Map<String, QueueHandler> queueHandlerMap;
+    abstract QueueHandler getQueueHandler(String queueName);
 
-    private final QueueDao queueDao;
+    abstract boolean addQueue(String queueName, boolean passive, boolean durable, boolean autoDelete,
+                              FieldTable arguments)
+            throws BrokerException;
 
-    private final QueueHandlerFactory queueHandlerFactory;
+    abstract int removeQueue(String queueName, boolean ifUnused, boolean ifEmpty) throws BrokerException,
+            ValidationException,
+            ResourceNotFoundException;
 
-    private final QueueRegistryEventPublisher queueRegistryEventPublisher;
-
-    public QueueRegistry(QueueDao queueDao,
-                         QueueHandlerFactory queueHandlerFactory,
-                         QueueRegistryEventPublisher queueRegistryEventPublisher) throws BrokerException {
-        this.queueHandlerMap = new HashMap<>();
-        this.queueDao = queueDao;
-        this.queueHandlerFactory = queueHandlerFactory;
-        this.queueRegistryEventPublisher = queueRegistryEventPublisher;
-        retrieveQueuesFromDao();
-    }
-
-
-    QueueHandler getQueueHandler(String queueName) {
-        return queueHandlerMap.get(queueName);
-    }
-
-    boolean addQueue(String queueName, boolean passive, boolean durable, boolean autoDelete, FieldTable arguments)
-            throws BrokerException {
-        QueueHandler queueHandler = queueHandlerMap.get(queueName);
-
-        if (passive) {
-            if (Objects.isNull(queueHandler)) {
-                throw new BrokerException("Queue [ " + queueName + " ] doesn't exists. Passive parameter "
-                                                  + "is set, hence not creating the queue.");
-            } else {
-                return false;
-            }
-        } else {
-            if (Objects.isNull(queueHandler)) {
-                if (durable) {
-                    queueHandler = queueHandlerFactory.createDurableQueueHandler(queueName, autoDelete, arguments);
-                    queueDao.persist(queueHandler.getUnmodifiableQueue());
-                } else {
-                    queueHandler = queueHandlerFactory.createNonDurableQueueHandler(queueName, autoDelete, arguments);
-                }
-                queueHandlerMap.put(queueName, queueHandler);
-                queueRegistryEventPublisher.publishQueueEvent("queue.created", queueHandler);
-                return true;
-            } else if (queueHandler.getUnmodifiableQueue().isDurable() != durable
-                       || queueHandler.getUnmodifiableQueue().isAutoDelete() != autoDelete) {
-                throw new BrokerException(
-                        "Existing queue [ " + queueName + " ] does not match given parameters.");
-            } else {
-                return false;
-            }
-        }
-    }
-
-    int removeQueue(String queueName, boolean ifUnused, boolean ifEmpty) throws BrokerException,
-                                                                                ValidationException,
-                                                                                ResourceNotFoundException {
-        QueueHandler queueHandler = queueHandlerMap.get(queueName);
-        if (queueHandler == null) {
-            throw new ResourceNotFoundException("Queue [ " + queueName + " ] Not found");
-        }
-
-        if (ifUnused && !queueHandler.isUnused()) {
-            throw new ValidationException("Cannot delete queue. Queue [ " + queueName
-                    + " ] has active consumers and the ifUnused parameter is set.");
-        } else if (ifEmpty && !queueHandler.isEmpty()) {
-            throw new ValidationException("Cannot delete queue. Queue [ " + queueName
-                    + " ] is not empty and the ifEmpty parameter is set.");
-        } else {
-            queueHandlerMap.remove(queueName);
-            queueDao.delete(queueHandler.getUnmodifiableQueue());
-            queueRegistryEventPublisher.publishQueueEvent("queue.deleted", queueHandler);
-            return queueHandler.releaseResources();
-        }
-    }
-
-    private void retrieveQueuesFromDao() throws BrokerException {
-            queueDao.retrieveAll((name) -> {
-                QueueHandler handler = queueHandlerFactory.createDurableQueueHandler(name, false,
-                        null);
-                queueHandlerMap.putIfAbsent(name, handler);
-            });
-    }
-
-    public Collection<QueueHandler> getAllQueues() {
-        return queueHandlerMap.values();
-    }
+    abstract Collection<QueueHandler> getAllQueues();
 
     /**
      * Method to reload queues on becoming the active node.
      *
      * @throws BrokerException if an error occurs loading messages from the database
      */
-    void reloadQueuesOnBecomingActive() throws BrokerException {
-        queueHandlerMap.clear();
-        retrieveQueuesFromDao();
-    }
-
-    /**
-     * Represents the event publisher handler in resource QueueRegistry.
-     */
-     interface QueueRegistryEventPublisher {
-        void publishQueueEvent(String eventType, QueueHandler queueHandler);
-    }
-
-    /**
-     * Default implementation of {@link QueueRegistryEventPublisher}.
-     */
-     static class DefaultQueueRegistryEventPublisher implements QueueRegistryEventPublisher {
-
-        private EventSync eventSync;
-         DefaultQueueRegistryEventPublisher(EventSync eventSync) {
-            this.eventSync = eventSync;
-        }
-
-        public void publishQueueEvent(String eventType, QueueHandler queueHandler) {
-            Map<String, String> properties = new HashMap<>();
-            String queueName = queueHandler.getUnmodifiableQueue().getName();
-            String isAutoDelete = String.valueOf(queueHandler.getUnmodifiableQueue().isAutoDelete());
-            String isDurable = String.valueOf(queueHandler.getUnmodifiableQueue().isDurable());
-            properties.put("queueName", queueName);
-            properties.put("autoDelete", isAutoDelete);
-            properties.put("durable", isDurable);
-            properties.put("messageCount", String.valueOf(queueHandler.size()));
-            eventSync.publish(eventType, properties);
-        }
-    }
-
-    /**
-     * Default implementation of {@link QueueRegistryEventPublisher}.
-     */
-     static class NullQueueRegistryEventPublisher implements QueueRegistryEventPublisher {
-
-        @Override
-        public void publishQueueEvent(String eventType, QueueHandler queueHandler) {
-            //Ignore
-        }
-    }
+    abstract void reloadQueuesOnBecomingActive() throws BrokerException;
 }
