+++ modules/broker-core/src/main/java/io/ballerina/messaging/broker/core/transaction/Branch.java
@@ -27,7 +27,6 @@ import io.ballerina.messaging.broker.core.Message;
 import io.ballerina.messaging.broker.core.QueueHandler;
 import io.ballerina.messaging.broker.core.store.MessageStore;
 import io.ballerina.messaging.broker.core.util.MessageTracer;
-
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
